requirements.txt
Django==1.4
South==0.7.5
django-registration==0.8
requests==0.13.6
django-email-as-username==1.6.2
django-compressor==1.2
